requirements_CoLMbo.txt
datasets==3.2.0
huggingface_hub==0.27.1
# importlib_resources==5.12.0
importlib_resources==6.4.0
librosa==0.10.0.post2
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.1
Pillow==11.3.0
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.1
soundfile==0.12.1
speechbrain==1.0.3
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
transformers==4.47.1
